update to v1.0!

+++ tensorflow/core/common_runtime/process_util.h
@@ -1,4 +1,4 @@
-/* Copyright 2016 Google Inc. All Rights Reserved.
+/* Copyright 2016 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -35,7 +35,7 @@ void SchedClosure(std::function<void()> closure);
 
 // Schedule "closure" after the given number of microseconds in the
 // fixed-size ThreadPool used for non-blocking compute tasks.
-void SchedNonBlockingClosureAfter(int micros, std::function<void()> closure);
+void SchedNonBlockingClosureAfter(int64 micros, std::function<void()> closure);
 
 }  // namespace tensorflow
